service done: for each add() checked if the id corresponds to another object and made sure to insert the created objects into the specific lists in entity

+++ src/main/java/com/example/movie_management_system/service/SeatService.java
@@ -2,13 +2,16 @@ package com.example.movie_management_system.service;
 
 import com.example.movie_management_system.model.Hall;
 import com.example.movie_management_system.model.Seat;
+import com.example.movie_management_system.model.Ticket;
 import com.example.movie_management_system.repository.HallRepositoryInFile;
 import com.example.movie_management_system.repository.SeatRepositoryInFile;
 import com.example.movie_management_system.repository.deprecated.SeatRepositoryInMemory;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -23,25 +26,39 @@ public class SeatService {
         this.seatRepository = seatRepository;
         this.hallService = hallService;
     }
-
+    @Transactional
     public void add(String hallId, String seatRow, String seatColumn ) {
         String id = UUID.randomUUID().toString();
         Seat seat = new Seat(id, hallId, seatRow, seatColumn);
-        seatRepository.add(seat);
+        if(hallService.findById(hallId).isPresent()) {
+            seatRepository.add(seat);
+            hallService.addSeat(hallId, seat);
+        }
     }
-
+    @Transactional
     public boolean update(String id, String hallId, String seatRow, String seatColumn) {
         Seat seat = new Seat(id, hallId, seatRow, seatColumn);
-        return seatRepository.update(id, seat);
+        hallService.updateSeat(hallId, id, seat);
+        seatRepository.update(id, seat);
+        return true;
+    }
+
+    @Transactional
+    public boolean remove(String id) {
+        Optional<Seat> seat = seatRepository.findById(id);
+        if (!seat.isPresent()) {
+            throw new NoSuchElementException("Seat with id " + id + " not found");
+        }
+        hallService.removeSeat(seat.get().getHallId(), id);
+        seatRepository.remove(id);
+
+        return true;
     }
 
     public List<Hall> getAvailableHalls(){
         return hallService.getAll();
     }
 
-    public void remove(String id) {
-        seatRepository.remove(id);
-    }
     public List<Seat> getAll() {
         return seatRepository.getAll();
     }
